requirements.txt
fastmcp>=0.1.0
fastapi>=0.103.1
uvicorn>=0.23.2
httpx>=0.24.1
pydantic>=2.3.0
starlette>=0.27.0
jinja2>=3.1.6